fetching star output logs to dataflow

+++ igf_data/utils/tools/star_utils.py
@@ -48,7 +48,7 @@ class Star_utils:
     :param two_pass_mode: Run two-pass mode of star, default True
     :param dry_run: A toggle forreturning the star cmd without actual run, default False
     :param star_patameters: A dictionary of star parameters, default encode parameters
-    :returns: A genomic_bam and a transcriptomic bam and star commandline
+    :returns: A genomic_bam and a transcriptomic bam,log file and star commandline
     '''
     try:
       temp_dir=get_temp_dir(use_ephemeral_space=True)                           # get a temp dir
@@ -112,28 +112,35 @@ class Star_utils:
       subprocess.check_call(star_cmd,shell=False)
       genomic_bam=''
       transcriptomic_bam=''
-      genomic_bam_pattern=re.compile(r'\S+sortedByCoord.out.bam$')              # pattern for genomic bam
-      transcriptomic_bam_pattern=re.compile(r'\S+toTranscriptome.out.bam$')     # pattern for transcriptomic bam
+      star_log_file=''
+      genomic_bam_pattern=re.compile(r'\S+sortedByCoord\.out\.bam$')            # pattern for genomic bam
+      transcriptomic_bam_pattern=re.compile(r'\S+toTranscriptome\.out\.bam$')   # pattern for transcriptomic bam
+      star_log_pattern=re.compile(r'\S+Log\.final\.out$')                       # pattern for star final log
       for file in os.listdir(temp_dir):
         if fnmatch.fnmatch(file, '*.bam'):
           source_path=os.path.join(temp_dir,file)
           destinationa_path=os.path.join(self.output_dir,file)
           copy_local_file(source_path=source_path,
                           destinationa_path=destinationa_path)                  # copying bams to output dir
-          if re.match(genomic_bam_pattern,
-                      os.path.basename(destinationa_path)):
+          if re.match(genomic_bam_pattern,file):
             genomic_bam=destinationa_path
 
-          if re.match(transcriptomic_bam_pattern,
-                      os.path.basename(destinationa_path)):
+          if re.match(transcriptomic_bam_pattern,file):
             transcriptomic_bam=destinationa_path
 
+        if re.match(star_log_pattern,file):
+          star_log_file=os.path.join(self.output_dir,file)
+          copy_local_file(source_path=os.path.join(temp_dir,file),
+                          destinationa_path=star_log_file)                      # copy star log file
+          
+
       if genomic_bam == '' or \
-         transcriptomic_bam == '':
-        raise ValueError('Star output bam files not found')
+         transcriptomic_bam == '' or \
+         star_log_file=='':
+        raise ValueError('Star output bam files or log file not found')
 
       remove_dir(temp_dir)                                                      # removing temp run dir
-      return genomic_bam,transcriptomic_bam,star_cmd
+      return genomic_bam,transcriptomic_bam,star_log_file,star_cmd
     except:
       if os.path.exists(temp_dir):
         remove_dir(temp_dir)
